+++ cms/src/app/index.tsx
@@ -1,24 +1,13 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 
-// import { createBrowserRouter } from "react-router";
 import App from "./App";
-// import routes from "../routes/routes";
 import "./index.css";
 import "./reset.styles.css";
-import rootRoute from "../routes/__root";
-import homeRoute from "../routes/homeRoute";
 import { createRouter, RouterProvider } from "@tanstack/react-router";
-
-// const router = createBrowserRouter(routes);
-const routeTree = rootRoute.addChildren([homeRoute]);
-const router = createRouter({ routeTree, InnerWrap: App });
-
-declare module "@tanstack/react-router" {
-  interface Register {
-    router: typeof router;
-  }
-}
+// import AuthProvider from "../hooks/useAuth";
+import UserProvider from "../hooks/useUser";
+import router from "../config/router.config";
 
 const rootElement: HTMLElement | null = document.getElementById("root");
 if (!rootElement) throw new Error(`Failed to find the root element`);
@@ -28,3 +17,20 @@ ReactDOM.createRoot(rootElement).render(
     <RouterProvider router={router} />
   </React.StrictMode>
 );
+
+/* ReactDOM.createRoot(rootElement).render(
+  <React.StrictMode>
+    <AuthProvider>
+      <UserProvider>
+        <RouterProvider router={router} />
+      </UserProvider>
+    </AuthProvider>
+  </React.StrictMode>
+); */
+
+/* ReactDOM.createRoot(rootElement).render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
+ */