+++ cms/src/components/header/Header.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect } from "react";
-import { Link, useLocation, useNavigate } from "react-router";
 
 import NavAnchor from "../navAnchor/NavAnchor";
 import styles from "./Header.module.css";
 import { useAuth } from "../../hooks/useAuth";
 import { usePrevLocation } from "../../hooks/usePrevLocation";
 import { useUserData } from "../../hooks/useUser";
+import { Link, useLocation, useNavigate } from "@tanstack/react-router";
 
 const Header: React.FC = () => {
   const { accessToken, isAuthenticated, logout } = useAuth();
@@ -30,7 +30,7 @@ const Header: React.FC = () => {
     console.log("from:", from);
     console.groupEnd();
     await logout();
-    navigate(from), { replace: true };
+    navigate({ to: from, replace: true });
   };
 
   return (
@@ -80,6 +80,37 @@ const Header: React.FC = () => {
       </nav>
     </header>
   );
+  /* return (
+    <header>
+      <nav>
+        <ul className={styles["nav-left"]}>
+          <li>
+            <Link to="/">
+              <h1>Project Name</h1>
+            </Link>
+          </li>
+        </ul>
+
+        <ul className={styles["nav-right"]}>
+          <li>
+            <NavAnchor pathname="/" textContent="home" />
+          </li>
+
+          <li>
+            <NavAnchor pathname="/faq" textContent="faq" />
+          </li>
+
+          <li>
+            <NavAnchor pathname="/login" textContent="login" />
+          </li>
+
+          <li>
+            <NavAnchor pathname="/signup" textContent="sign up" />
+          </li>
+        </ul>
+      </nav>
+    </header>
+  ); */
 };
 
 export default Header;