+++ cms/src/components/navAnchor/NavAnchor.tsx
@@ -1,5 +1,4 @@
-import React from "react";
-import { NavLink } from "react-router";
+import { Link } from "@tanstack/react-router";
 
 /* interface Props<T = void> {
   pathname: string;
@@ -15,7 +14,7 @@ interface Props<T = void> {
  * https://stackoverflow.com/questions/53958028/how-to-use-generics-in-props-in-react-in-a-functional-component/66052574#66052574
  */
 const NavAnchor = <T,>({ pathname, textContent }: Props<T>) => {
-  return <NavLink to={`${pathname}`}>{textContent}</NavLink>;
+  return <Link to={`${pathname}`}>{textContent}</Link>;
 };
 
 export default NavAnchor;